+++ lib/main.dart
@@ -1,10 +1,12 @@
 
 import 'package:flutter/material.dart';
+import 'package:sg_mobile/service/service_locator.dart';
 import 'package:sg_mobile/widgets/login.dart';
 import 'package:sg_mobile/widgets/square_game_app_bar.dart';
 
 
 void main() {
+  setupServiceLocator();
   runApp(const MyApp());
 }
