+++ pubspec.yaml
@@ -36,6 +36,7 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
   http: ^0.13.6
+  get_it: ^7.6.0
 
 dev_dependencies:
   flutter_test:
